images updated

+++ admin/about/AboutImageChange.php
@@ -42,7 +42,11 @@ else{
          setcookie('about', '', time()-3600, '/');
            echo "Successful,";
            echo $filenamenew;
-           unlink($previousFileLocation);
+           if(file_exists($previousFileLocation)){
+            chmod($previousFileLocation, 0644);
+            unlink($previousFileLocation);
+           }
+           
            
         }
         else{
